mudarmatricula done

+++ app/src/main/java/com/example/parkingsystem/model/post/Utilizador.kt
@@ -1,5 +1,5 @@
 package com.example.parkingsystem.model.post
 
 data class UtilizadorPost (
-    val idUtilizador: Long
+    val idUtilizador: Int
 )
